+++ tests/testthat/test-aaa.R
@@ -1,0 +1,3 @@
+test_that("global tracedump turned off at start", {
+  expect_true(isFALSE(global_ggtrace_state()))
+})